fix: remaining int64 migration cleanup
- NdArray.ToString.cs: Change loop variable from int to long for size iteration
- np.roll.cs: Remove unnecessary (int) casts since Slice accepts long natively

+++ src/NumSharp.Core/Casting/NdArray.ToString.cs
@@ -69,7 +69,7 @@ namespace NumSharp
             var size = shape[0];
             s.Append("[");
 
-            for (int i = 0; i < size; i++)
+            for (long i = 0; i < size; i++)
             {
                 var n_minus_one_dim_slice = this[i];
                 n_minus_one_dim_slice.PrettyPrint(s, flat);
